fix pid check

+++ backend/onyx/background/celery/tasks/indexing/tasks.py
@@ -104,15 +104,17 @@ class IndexingCallback(IndexingHeartbeatInterface):
         return False
 
     def progress(self, tag: str, amount: int) -> None:
-        now = time.monotonic()
-        if now - self.last_parent_check > IndexingCallback.PARENT_CHECK_INTERVAL:
-            try:
-                # this is unintuitive, but it checks if the parent pid is still running
-                os.kill(self.parent_pid, 0)
-            except Exception:
-                logger.exception("IndexingCallback - parent pid check exceptioned")
-                raise
-            self.last_parent_check = now
+        if self.parent_pid:
+            # check if the parent pid is alive so we aren't running as a zombie
+            now = time.monotonic()
+            if now - self.last_parent_check > IndexingCallback.PARENT_CHECK_INTERVAL:
+                try:
+                    # this is unintuitive, but it checks if the parent pid is still running
+                    os.kill(self.parent_pid, 0)
+                except Exception:
+                    logger.exception("IndexingCallback - parent pid check exceptioned")
+                    raise
+                self.last_parent_check = now
 
         try:
             self.redis_lock.reacquire()
